+++ .umirc.js
@@ -31,6 +31,10 @@ export default defineConfig({
       path: '/playground',
     },
     {
+      title: '常见问题',
+      path: '/faq',
+    },
+    {
       title: '周边工具',
       path: '/tools',
       children: [
@@ -89,7 +93,7 @@ export default defineConfig({
   sitemap: {
     hostname: 'https://xrender.fun',
   },
-
+  runtimePublicPath: true,
   chainWebpack(config, { webpack }) {
     config.plugin('monaco-editor').use(MonacoWebpackPlugin);
   },