+++ packages/form-render/src/form-render-core/src/core/RenderChildren/RenderList/CardList.js
@@ -22,8 +22,12 @@ const CardList = ({
   displayType,
   getFieldsProps,
 }) => {
-  const { props = {}, itemProps } = schema;
-  const { methods } = useTools();
+  const { props = {}, itemProps, min = 0, max = 9999 } = schema;
+  const { methods, widgets } = useTools();
+
+  const CustomAddBtn = widgets[schema['add-widget']];
+
+  const AddWidget = CustomAddBtn || Button;
 
   let addBtnProps = {
     type: 'dashed',
@@ -44,6 +48,8 @@ const CardList = ({
     delConfirmProps = { ...delConfirmProps, ...props.delConfirmProps };
   }
 
+  addBtnProps.onClick = addItem;
+
   return (
     <>
       <div className="fr-card-list">
@@ -57,7 +63,10 @@ const CardList = ({
               }`}
               key={idx}
             >
-              <div className="fr-card-index">{idx + 1}</div>
+              <div className="fr-card-index">
+                {/* <>{props.prefix}</> */}
+                {idx + 1}
+              </div>
               <Core {...fieldsProps} />
 
               <div direction="horizontal" className="fr-card-toolbar">
@@ -79,7 +88,7 @@ const CardList = ({
                     onClick={() => copyItem(idx)}
                   />
                 )}
-                {!props.hideDelete && (
+                {!props.hideDelete && displayList.length > min && (
                   <Popconfirm
                     onConfirm={() => {
                       if (
@@ -88,7 +97,7 @@ const CardList = ({
                       ) {
                         const cb = methods[props.onConfirm];
                         if (typeof cb === 'function') {
-                          const result = cb();
+                          const result = cb(item, idx);
                           if (!result) {
                             return;
                           }
@@ -107,7 +116,9 @@ const CardList = ({
         })}
       </div>
       <div style={{ marginTop: displayList.length > 0 ? 0 : 8 }}>
-        {!props.hideAdd && <Button onClick={addItem} {...addBtnProps} />}
+        {!props.hideAdd && displayList.length < max && (
+          <AddWidget {...addBtnProps} />
+        )}
         {Array.isArray(props.buttons)
           ? props.buttons.map((item, idx) => {
               const { callback, text, html } = item;