+++ packages/form-render/src/form-render-core/src/core/RenderChildren/RenderList/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import { get } from 'lodash-es';
+import { get, isFunction } from 'lodash-es';
 import React from 'react';
 import { useStore, useTools } from '../../../hooks';
 import {
@@ -24,8 +24,7 @@ const RenderList = ({
   displayType,
 }) => {
   const { formData, flatten } = useStore();
-  const { onItemChange, removeTouched, methods } = useTools();
-
+  const { onItemChange, removeTouched, methods, layoutWidgets } = useTools();
   const { props = {} } = schema;
 
   let renderWidget = 'list';
@@ -129,7 +128,7 @@ const RenderList = ({
     };
   };
 
-  const displayProps = {
+  let displayProps = {
     displayList,
     changeList,
     schema,
@@ -147,6 +146,23 @@ const RenderList = ({
     displayType,
     getFieldsProps,
   };
+
+  // 外部定义：添加按钮事件
+  const onAdd = methods[props.onAdd];
+  if (isFunction(onAdd)) {
+    displayProps.addItem = () => onAdd(addItem, { schema });
+  }
+
+  // 外部定义：删除按钮事件
+  const onRemove = methods[props.onRemove];
+  if (isFunction(onRemove)) {
+    displayProps.deleteItem = (idx) => onRemove(() => deleteItem(idx), { schema })
+  }
+
+  if (layoutWidgets && layoutWidgets[renderWidget]) {
+    const Component = layoutWidgets[renderWidget];
+    return <Component {...displayProps} />;
+  }
 
   switch (renderWidget) {
     case 'list1':