+++ packages/form-render/src/form-render-core/src/core/RenderField/Title.js
@@ -1,25 +1,47 @@
 import { Tooltip } from 'antd';
 import React from 'react';
-import { useStore2 } from '../../hooks';
+import { useStore2, useTools } from '../../hooks';
 import { isCheckBoxType } from '../../utils';
 
 const Description = ({ displayType, schema }) => {
-  const { description, descType } = schema;
-  if (!description) return null;
+  const { description, descType, descWidget } = schema;
+  if (!description && !descWidget) return null;
+
+  const { widgets } = useTools();
+
+  const _description =
+    typeof description === 'string' && /(^<|\/>)/.test(description) ? (
+      <div dangerouslySetInnerHTML={{ __html: description }} />
+    ) : (
+      description
+    );
+  const RenderDesc = () => {
+    const Widget = widgets[schema.descWidget];
+    if (Widget) {
+      return <Widget schema={schema} />;
+    }
+    return null;
+  };
 
   switch (displayType) {
     case 'row':
+      if (descType === 'widget') {
+        return <RenderDesc />;
+      }
       return (
-        <Tooltip title={description}>
+        <Tooltip title={_description}>
           <i className="fr-tooltip-icon" />
         </Tooltip>
       );
     case 'inline':
       return null;
     default:
+      if (descType === 'widget') {
+        return <RenderDesc />;
+      }
       if (descType === 'icon') {
         return (
-          <Tooltip title={description}>
+          <Tooltip title={_description}>
             <i className="fr-tooltip-icon" />
           </Tooltip>
         );
@@ -55,35 +77,37 @@ const Title = ({
     });
   }
 
-  const requiredMark = typeof schemaRequiredMark === 'undefined' ? globalRequiredMark : schemaRequiredMark;
+  const requiredMark =
+    typeof schemaRequiredMark === 'undefined'
+      ? globalRequiredMark
+      : schemaRequiredMark;
 
   // 左侧的的 * 号提示
   let TitleRequiredMark = null;
   // 左侧的 option 提示
   let TitleTextMark = null;
 
-  if(required) {
+  if (required) {
     /**
      * ant-design requiredMark 实现
      * https://ant.design/components/form-cn/
      */
-    if(requiredMark !== false && requiredMark !== 'optional') {
+    if (requiredMark !== false && requiredMark !== 'optional') {
       TitleRequiredMark = <span className="fr-label-required"> *</span>;
       TitleTextMark = null;
     }
   } else {
-    if(requiredMark === 'optional') {
+    if (requiredMark === 'optional') {
       TitleRequiredMark = null;
       TitleTextMark = <span className="fr-label-required-text">（可选）</span>;
     }
   }
 
   // requiredMark 为 false 不展示必填符号
-  if(requiredMark === false) {
+  if (requiredMark === false) {
     TitleRequiredMark = null;
     TitleTextMark = null;
   }
-
 
   return (
     <div className={labelClass} style={labelStyle}>